ex26.py "TEST"
Here's the test, I believe that I've covered everything. I will not only look over it today, but I will review it tomorrow as well. In the book, Zed claimed that this test would be "challenging". I actually found it to be rather simple. Hopefully, I got most everything right!